added forecasting and updated chart
