+++ composer.json
@@ -10,7 +10,8 @@
         "guzzlehttp/guzzle": "^7.2",
         "laravel/framework": "^10.10",
         "laravel/sanctum": "^3.3",
-        "laravel/tinker": "^2.8"
+        "laravel/tinker": "^2.8",
+        "league/csv": "^9.15"
     },
     "require-dev": {
         "fakerphp/faker": "^1.9.1",
