+++ resources/views/admin/admin_dashboard.blade.php
@@ -76,10 +76,6 @@ License: For each use you must have a valid license purchased only from above li
 	<script src="{{ asset('../assets/js/template.js')}}"></script>
 	<!-- endinject -->
 
-	<!-- Custom js for this page -->
-  <script src="{{ asset('../assets/js/dashboard-light.js')}}"></script>					{{--JAVASCRIPT DATA OF MONITORING GRAPH --}}
-	<!-- End custom js for this page -->
-
     {{-- PM2.5 Chart--}}
   	<script src="{{ asset('../assets/pm25.js')}}"></script>
     {{-- PM10 Chart--}}
