+++ resources/views/admin/index.blade.php
@@ -77,8 +77,6 @@
 
               @include('charts.monitoring.o3')
 
-              @include('charts.monitoring.avg')
-
             </div>
           </div>
         </div>
