+++ resources/views/charts/forecastingpm10.blade.php
@@ -6,7 +6,7 @@
         <h5 class="card-title mb-0">PM10</h5>
         <div class="row align-items-start">
           <div class="col-md-7">
-            <p class="text-muted tx-13 mb-3 mb-md-0">Forecasting PM10 using Random Forest Regressor Algorithm</p>
+            <p class="text-muted tx-13 mb-3 mb-md-0">Forecasting PM10 Concentration for Yr. 2024</p>
           </div>
           <div class="col-md-5 d-flex justify-content-md-end">
           </div>
