--- resources/views/charts/monitoring/avg.blade.php
@@ -1,15 +1,0 @@
-<div class="col-md-6 col-sm-4 grid-margin stretch-card">
-    <div class="card">
-      <div class="card-body">
-        <div class="d-flex justify-content-between align-items-baseline">
-        </div>
-        <div class="row">
-          <div class="d-flex align-items-baseline justify-content-between">
-            <h6 class="d-flex align-items-center">Avg.</h6>
-          </div>
-            <div id="Avg" class="mt-md-3 mt-xl-0"></div>
-        </div>
-
-      </div>
-    </div>
-  </div>
