+++ resources/views/user/index.blade.php
@@ -70,8 +70,6 @@
 
               @include('charts.monitoring.o3')
 
-              @include('charts.monitoring.avg')
-
             </div>
           </div>
         </div>
