+++ storage/Data/average_pm10_per_day_2024.csv
@@ -1,0 +1,367 @@
+Date,Average PM10
+1/1/2024,57.94125
+1/2/2024,30.70458333
+1/3/2024,41.18708333
+1/4/2024,43.5125
+1/5/2024,44.2025
+1/6/2024,54.27541667
+1/7/2024,51.95458333
+1/8/2024,55.64166667
+1/9/2024,40.42333333
+1/10/2024,55.37041667
+1/11/2024,50.03708333
+1/12/2024,53.46916667
+1/13/2024,32.42708333
+1/14/2024,40.06208333
+1/15/2024,38.73708333
+1/16/2024,46.68583333
+1/17/2024,50.50708333
+1/18/2024,64.2675
+1/19/2024,60.35166667
+1/20/2024,49.91166667
+1/21/2024,48.76458333
+1/22/2024,50.99
+1/23/2024,46.84916667
+1/24/2024,53.55083333
+1/25/2024,57.43041667
+1/26/2024,47.23916667
+1/27/2024,48.11166667
+1/28/2024,35.91375
+1/29/2024,40.73083333
+1/30/2024,29.94625
+1/31/2024,32.86625
+2/1/2024,46.36041667
+2/2/2024,58.735
+2/3/2024,63.1475
+2/4/2024,57.66541667
+2/5/2024,37.39791667
+2/6/2024,32.61041667
+2/7/2024,52.55541667
+2/8/2024,71.53666667
+2/9/2024,48.71416667
+2/10/2024,44.68916667
+2/11/2024,44.9875
+2/12/2024,53.39083333
+2/13/2024,36.42666667
+2/14/2024,54.11083333
+2/15/2024,42.88458333
+2/16/2024,41.89125
+2/17/2024,48.28166667
+2/18/2024,54.40625
+2/19/2024,50.92625
+2/20/2024,48.59708333
+2/21/2024,57.60416667
+2/22/2024,51.75375
+2/23/2024,55.92458333
+2/24/2024,53.88916667
+2/25/2024,49.085
+2/26/2024,47.09708333
+2/27/2024,48.25375
+2/28/2024,33.65375
+2/29/2024,38.19041667
+3/1/2024,46.59291667
+3/2/2024,57.11125
+3/3/2024,46.33875
+3/4/2024,52.62708333
+3/5/2024,42.43875
+3/6/2024,38.05083333
+3/7/2024,56.16416667
+3/8/2024,62.66916667
+3/9/2024,47.93416667
+3/10/2024,47.20708333
+3/11/2024,59.96708333
+3/12/2024,52.225
+3/13/2024,33.77041667
+3/14/2024,39.36708333
+3/15/2024,39.25
+3/16/2024,42.88625
+3/17/2024,49.75583333
+3/18/2024,59.0125
+3/19/2024,52.12291667
+3/20/2024,51.83125
+3/21/2024,43.715
+3/22/2024,44.36708333
+3/23/2024,42.62
+3/24/2024,50.41416667
+3/25/2024,59.36541667
+3/26/2024,42.32541667
+3/27/2024,41.25
+3/28/2024,35.3825
+3/29/2024,38.18
+3/30/2024,34.79041667
+3/31/2024,34.20541667
+4/1/2024,37.78375
+4/2/2024,41.74875
+4/3/2024,46.34
+4/4/2024,51.13833333
+4/5/2024,44.48625
+4/6/2024,44.2775
+4/7/2024,57.02958333
+4/8/2024,57.93916667
+4/9/2024,46.85041667
+4/10/2024,42.94
+4/11/2024,43.23041667
+4/12/2024,50.31416667
+4/13/2024,41.65416667
+4/14/2024,44.46125
+4/15/2024,43.06083333
+4/16/2024,46.3425
+4/17/2024,50.37625
+4/18/2024,57.30958333
+4/19/2024,56.7475
+4/20/2024,49.61625
+4/21/2024,50.47833333
+4/22/2024,52.98416667
+4/23/2024,55.20958333
+4/24/2024,53.78
+4/25/2024,57.3225
+4/26/2024,46.55458333
+4/27/2024,42.48666667
+4/28/2024,39.74375
+4/29/2024,49.99791667
+4/30/2024,63.71166667
+5/1/2024,41.44583333
+5/2/2024,38.03041667
+5/3/2024,33.93375
+5/4/2024,30.16541667
+5/5/2024,29.01958333
+5/6/2024,28.59125
+5/7/2024,43.19083333
+5/8/2024,54.075
+5/9/2024,40.54541667
+5/10/2024,41.00458333
+5/11/2024,33.07625
+5/12/2024,32.91208333
+5/13/2024,31.79541667
+5/14/2024,34.3
+5/15/2024,36.48
+5/16/2024,30.44625
+5/17/2024,38.10375
+5/18/2024,38.90166667
+5/19/2024,39.11875
+5/20/2024,48.56625
+5/21/2024,38.94375
+5/22/2024,41.46666667
+5/23/2024,45.72125
+5/24/2024,65.4775
+5/25/2024,63.01833333
+5/26/2024,56.37666667
+5/27/2024,52.39416667
+5/28/2024,49.20458333
+5/29/2024,59.24
+5/30/2024,72.42916667
+5/31/2024,68.52166667
+6/1/2024,66.32666667
+6/2/2024,46.10708333
+6/3/2024,48.27666667
+6/4/2024,49.63958333
+6/5/2024,42.84041667
+6/6/2024,38.5225
+6/7/2024,44.03541667
+6/8/2024,51.52916667
+6/9/2024,46.25458333
+6/10/2024,52.6875
+6/11/2024,56.63416667
+6/12/2024,49.50208333
+6/13/2024,45.9925
+6/14/2024,45.17125
+6/15/2024,43.77666667
+6/16/2024,42.91125
+6/17/2024,45.59083333
+6/18/2024,52.02875
+6/19/2024,50.08541667
+6/20/2024,53.91375
+6/21/2024,46.48666667
+6/22/2024,47.75375
+6/23/2024,48.74875
+6/24/2024,64.34875
+6/25/2024,62.25541667
+6/26/2024,55.9475
+6/27/2024,52.32291667
+6/28/2024,49.09291667
+6/29/2024,58.58708333
+6/30/2024,71.86333333
+7/1/2024,46.89458333
+7/2/2024,42.29666667
+7/3/2024,42.1025
+7/4/2024,29.96458333
+7/5/2024,32.305
+7/6/2024,27.55916667
+7/7/2024,40.52375
+7/8/2024,38.90541667
+7/9/2024,38.74208333
+7/10/2024,27.77541667
+7/11/2024,42.24291667
+7/12/2024,39.31166667
+7/13/2024,37.49666667
+7/14/2024,42.23041667
+7/15/2024,44.86791667
+7/16/2024,33.51708333
+7/17/2024,50.76375
+7/18/2024,38.46458333
+7/19/2024,29.92625
+7/20/2024,28.50083333
+7/21/2024,24.14333333
+7/22/2024,25.35333333
+7/23/2024,28.70083333
+7/24/2024,27.64375
+7/25/2024,33.05291667
+7/26/2024,47.73416667
+7/27/2024,45.35375
+7/28/2024,37.06375
+7/29/2024,29.5725
+7/30/2024,53.67125
+7/31/2024,52.49708333
+8/1/2024,47.04583333
+8/2/2024,51.22916667
+8/3/2024,52.45416667
+8/4/2024,50.0475
+8/5/2024,50.5225
+8/6/2024,42.905
+8/7/2024,40.42875
+8/8/2024,36.5375
+8/9/2024,33.68666667
+8/10/2024,31.18583333
+8/11/2024,31.02458333
+8/12/2024,56.33625
+8/13/2024,65.59416667
+8/14/2024,60.51458333
+8/15/2024,49.44333333
+8/16/2024,45.48083333
+8/17/2024,41.95333333
+8/18/2024,45.43458333
+8/19/2024,41.66166667
+8/20/2024,41.38833333
+8/21/2024,52.365
+8/22/2024,45.5625
+8/23/2024,47.45333333
+8/24/2024,45.74291667
+8/25/2024,55.84541667
+8/26/2024,48.77041667
+8/27/2024,43.09125
+8/28/2024,50.46625
+8/29/2024,34.8
+8/30/2024,36.095
+8/31/2024,35.31708333
+9/1/2024,40.88458333
+9/2/2024,41.89541667
+9/3/2024,32.42583333
+9/4/2024,25.34791667
+9/5/2024,24.35375
+9/6/2024,35.62458333
+9/7/2024,34.315
+9/8/2024,21.89625
+9/9/2024,30.98541667
+9/10/2024,36.38208333
+9/11/2024,24.995
+9/12/2024,27.58708333
+9/13/2024,44.6075
+9/14/2024,49.74833333
+9/15/2024,50.89
+9/16/2024,43.08291667
+9/17/2024,37.46666667
+9/18/2024,38.34916667
+9/19/2024,35.24416667
+9/20/2024,40.75958333
+9/21/2024,39.71458333
+9/22/2024,27.33958333
+9/23/2024,32.55333333
+9/24/2024,40.96208333
+9/25/2024,43.73291667
+9/26/2024,43.85458333
+9/27/2024,29.08541667
+9/28/2024,36.27666667
+9/29/2024,32.88958333
+9/30/2024,31.94958333
+10/1/2024,38.86625
+10/2/2024,51.95166667
+10/3/2024,31.09333333
+10/4/2024,30.07291667
+10/5/2024,30.87166667
+10/6/2024,38.47833333
+10/7/2024,50.85875
+10/8/2024,71.28708333
+10/9/2024,36.55958333
+10/10/2024,30.95416667
+10/11/2024,44.1275
+10/12/2024,41.24625
+10/13/2024,31.25166667
+10/14/2024,32.93125
+10/15/2024,32.14208333
+10/16/2024,39.355
+10/17/2024,27.85333333
+10/18/2024,29.41708333
+10/19/2024,53.92666667
+10/20/2024,41.85125
+10/21/2024,46.42
+10/22/2024,44.95083333
+10/23/2024,35.79541667
+10/24/2024,29.90166667
+10/25/2024,35.70625
+10/26/2024,66.72875
+10/27/2024,59.78666667
+10/28/2024,40.3875
+10/29/2024,29.33791667
+10/30/2024,34.23583333
+10/31/2024,36.17833333
+11/1/2024,25.6275
+11/2/2024,31.20333333
+11/3/2024,36.47083333
+11/4/2024,38.15333333
+11/5/2024,33.45916667
+11/6/2024,32.27833333
+11/7/2024,39.53125
+11/8/2024,34.49166667
+11/9/2024,37.94208333
+11/10/2024,54.12416667
+11/11/2024,41.41666667
+11/12/2024,56.58083333
+11/13/2024,53.38875
+11/14/2024,59.72708333
+11/15/2024,50.20833333
+11/16/2024,47.48833333
+11/17/2024,35.39416667
+11/18/2024,47.03708333
+11/19/2024,51.17083333
+11/20/2024,49.26458333
+11/21/2024,37.13
+11/22/2024,55.64833333
+11/23/2024,53.99791667
+11/24/2024,48.91625
+11/25/2024,48.755
+11/26/2024,55.08541667
+11/27/2024,54.07791667
+11/28/2024,54.70208333
+11/29/2024,68.56625
+11/30/2024,55.44833333
+12/1/2024,39.08041667
+12/2/2024,52.77666667
+12/3/2024,60.04166667
+12/4/2024,62.49166667
+12/5/2024,63.08291667
+12/6/2024,76.88291667
+12/7/2024,65.06416667
+12/8/2024,46.15333333
+12/9/2024,37.0125
+12/10/2024,48.265
+12/11/2024,40.07041667
+12/12/2024,34.96833333
+12/13/2024,45.3775
+12/14/2024,65.13083333
+12/15/2024,68.89
+12/16/2024,57.44541667
+12/17/2024,37.50291667
+12/18/2024,34.32666667
+12/19/2024,36.54583333
+12/20/2024,46.66958333
+12/21/2024,50.09958333
+12/22/2024,52.78291667
+12/23/2024,59.93291667
+12/24/2024,49.99125
+12/25/2024,48.22291667
+12/26/2024,32.18083333
+12/27/2024,28.33375
+12/28/2024,42.35541667
+12/29/2024,45.2925
+12/30/2024,47.04666667
+12/31/2024,34.93
